+++ frontend/packages/admin-app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @haste-health/admin-app
 
+## 0.19.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @haste-health/components@0.19.0
+
 ## 0.19.1
 
 ### Patch Changes