+++ frontend/packages/smart-launch/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @haste-health/smart-launch
 
+## 0.7.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @haste-health/components@0.19.0
+
 ## 0.7.1
 
 ### Patch Changes